Simplify adding bower components to sass path
Make use of .bower.json file for each component.  This file specifies
the main directory, which is the file we will want to include.  Use this
path rather than trying to guess the path based on the component name
(this guesswork approach is what we were doing before. it worked but it
may not work for non uniform components)