Move assets over to use jekyll-assets
Makes use of sprockets and better cache busting.
Move assets to _assets.
Removes the need for hacky entry css file, that is actually sass but
with content on the top to inform jekyll.

This is a good move off of the built in jekyll assets as they will no
longer be included with jekyll by default in future releases.