Fix monospace vertical rhythm issue (for the millionth time)
For some reason, vertical rhythm is off for monospace fonts that
are not Courier. This only happens with a base font-size of 18px (1rem).
Example font: Inconsolata. Possibly a sub-pixel rounding error? Or
browser error rendering monospace fonts? Or problem with the font
itself?

Use Courier until this problem is solved.